requirements.txt
# Core Dependencies for ImageGeneratorLLM
# Python 3.13+ compatible

# PyTorch with MPS (Metal Performance Shaders) support for Apple Silicon
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Hugging Face libraries for model loading and inference
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# Image processing
Pillow>=10.1.0

# REST API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# CLI interface
click>=8.1.7

# Utilities
numpy>=1.26.0
pyyaml>=6.0.1
requests>=2.31.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Code quality
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
